ImgMath: refactored classes and interfaces into their own files.
Also added an static ImgMath.compute method to further
ease the use of the library.
