+++ src/main/java/net/imglib2/algorithm/math/abstractions/IBinaryFunction.java
@@ -1,0 +1,6 @@
+package net.imglib2.algorithm.math.abstractions;
+
+public interface IBinaryFunction extends IUnaryFunction
+{
+	public IFunction getSecond();
+}
