+++ src/main/java/net/imglib2/algorithm/math/abstractions/ITrinaryFunction.java
@@ -1,0 +1,6 @@
+package net.imglib2.algorithm.math.abstractions;
+
+public interface ITrinaryFunction extends IBinaryFunction
+{
+	public IFunction getThird();
+}
