+++ src/main/java/net/imglib2/algorithm/math/abstractions/IUnaryFunction.java
@@ -1,0 +1,6 @@
+package net.imglib2.algorithm.math.abstractions;
+
+public interface IUnaryFunction extends IFunction
+{
+	public IFunction getFirst();
+}
